Update file_storage.py
Changes
1. The import of models
Initializing the imports at the beginning of the document seems to create a circular import error. The situation with BaseModel for example is BaseModel imports FileStorage to save new instances, and FileStorage imports from BaseModel to ensure we are able to reload data from JSON file back to object attribute. This seems to create a circle of imports that never ends. So I localized the imports to only occur at a particular instance. For now I am ignoring the error ModuleError: Module not found after running python3 <file_name>.